+++ aioxmpp/testutils.py
@@ -106,6 +106,19 @@ def run_coroutine_with_peer(
     return local_future.result()
 
 
+class FilterMock(unittest.mock.Mock):
+    def __init__(self):
+        super().__init__([
+            "register",
+            "unregister",
+            "filter",
+        ])
+        self.context_register = unittest.mock.MagicMock([
+            "__enter__",
+            "__exit__",
+        ])
+
+
 class ConnectedClientMock(unittest.mock.Mock):
     on_stream_established = callbacks.Signal()
     on_stream_destroyed = callbacks.Signal()
@@ -130,6 +143,14 @@ class ConnectedClientMock(unittest.mock.Mock):
         self.stream_features = nonza.StreamFeatures()
         self.stream.on_message_received = callbacks.AdHocSignal()
         self.stream.on_presence_received = callbacks.AdHocSignal()
+        self.stream.app_inbound_message_filter = FilterMock()
+        self.stream.app_inbound_presence_filter = FilterMock()
+        self.stream.app_outbound_message_filter = FilterMock()
+        self.stream.app_outbound_presence_filter = FilterMock()
+        self.stream.service_inbound_message_filter = FilterMock()
+        self.stream.service_inbound_presence_filter = FilterMock()
+        self.stream.service_outbound_message_filter = FilterMock()
+        self.stream.service_outbound_presence_filter = FilterMock()
         self.stream.send_iq_and_wait_for_reply = CoroutineMock()
         self.stream.send = CoroutineMock()
         self.stream.enqueue_stanza = self.stream.enqueue